feat: add mixin for setting margins
This commit adds a mixin for setting margins to elements. The mixin takes two arguments,  and , which specify the left and right margins, respectively.

The mixin is used to set the margins of the body and div tags. The body tag has a margin of 10px on the left and right, and the div tag has a margin of 15px on the left and right.

This mixin can be reused to set the margins of any element. It is a good way to keep your Sass code DRY and organized.